# src/routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Viagem365",
    "description": "Projeto FuturoDev",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization",
      "description": "Token de Autenticação"
    }
  },
  "paths": {
    "/login/": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Faça login para autenticar",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "teste123@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "teste123"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/usuarios/": {
      "post": {
        "tags": [
          "Usuário"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Cadastrar um novo usuário",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Raul Medeiros"
                },
                "data_nascimento": {
                  "type": "string",
                  "example": "1996-12-15"
                },
                "sexo": {
                  "type": "string",
                  "example": "M"
                },
                "cpf": {
                  "type": "string",
                  "example": "01234567800"
                },
                "endereco": {
                  "type": "string",
                  "example": "Palhoça"
                },
                "email": {
                  "type": "string",
                  "example": "nome@dominio.com.br"
                },
                "password": {
                  "type": "string",
                  "example": "123456"
                }
              },
              "required": [
                "nome",
                "data_nascimento",
                "sexo",
                "cpf",
                "endereco",
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/destinos/": {
      "post": {
        "tags": [
          "Destino"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Cadastrar um novo destino",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Raul Medeiros"
                },
                "descricao": {
                  "type": "string",
                  "example": "2000-01-01"
                },
                "localidade": {
                  "type": "string",
                  "example": "Lagoa do Peri"
                },
                "cep": {
                  "type": "string",
                  "example": "11222999"
                }
              },
              "required": [
                "nome",
                "descricao",
                "localidade",
                "cep"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Destino"
        ],
        "description": "Esta rota requer autenticação com um token de usuário para listar todos os destinos.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/destinos/{local_id}": {
      "get": {
        "tags": [
          "Destino"
        ],
        "description": "Esta rota requer autenticação com um token de usuário para listar todos os destinos.",
        "parameters": [
          {
            "name": "local_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/destinos/{id}": {
      "put": {
        "tags": [
          "Destino"
        ],
        "description": "Esta rota requer autenticação com um token de usuário para listar todos os destinos.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "descricao": {
                  "example": "any"
                },
                "localidade": {
                  "example": "any"
                },
                "coordernadas": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Destino"
        ],
        "description": "Esta rota requer autenticação com um token de usuário para listar todos os destinos.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ]
}